added basic lineplot
