+++ frontend/src/stories/AltitudePlot.stories.tsx
@@ -1,43 +1,7 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import {XYChart, LineSeries, Axis } from '@visx/xychart';
+import AltitudePlot from '../components/AltitudePlot';
 
-
-type AltitudePlotProps = {
-  data: { x: number, y: number }[]
-}
-
-const AltitudePlot = (props: AltitudePlotProps) => {
-  console.log(props.data);
-  return (
-    <XYChart
-      xScale={{ type: 'band' }}
-      yScale={{ type: 'linear' }}
-      width={300}
-      height={300}
-    >
-      <Axis orientation="bottom" />
-      <Axis orientation="left" />
-      <LineSeries
-        data={[
-          { x: 0, y: 0 },
-          { x: 1, y: 1 },
-          { x: 2, y: 2 },
-          { x: 3, y: 3 },
-          { x: 4, y: 4 },
-          { x: 5, y: 5 },
-          { x: 6, y: 6 },
-          { x: 7, y: 7 },
-          { x: 8, y: 8 },
-          { x: 9, y: 9 },
-        ]}
-        xAccessor={(d: { x: number, y: number }) => d.x} 
-        yAccessor={(d: { x: number, y: number }) => d.y}
-        dataKey={"key1"}
-      />
-    </XYChart>
-  );
-}
 
 
 const meta = {
@@ -58,10 +22,19 @@ const meta = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
+
+const fakeData = [
+  {x: 0, y: 0},
+  {x: 1, y: -1},
+  {x: 2, y: 12},
+  {x: 3, y: 3},
+  {x: 4, y: -4},
+  {x: 5, y: 5},
+  {x: 6, y: 16},
+];
+
 export const Primary: Story = {
   args: {
-    data: [
-      {x: 0, y: 0},
-    ]
+    data: fakeData
   }
 };
